requirements.txt
boto3
loguru
pandas
pytz
pyyaml
requests
slackclient
